+++ src/main/java/windeath44/server/memorial/domain/presentation/controller/MemorialPullRequestController.java
@@ -1,0 +1,25 @@
+package windeath44.server.memorial.domain.presentation.controller;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import windeath44.server.memorial.domain.presentation.dto.ResponseDto;
+import windeath44.server.memorial.domain.presentation.dto.request.MemorialPullRequestDto;
+import windeath44.server.memorial.domain.service.MemorialPullRequestService;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/memorials")
+public class MemorialPullRequestController {
+  MemorialPullRequestService memorialPullRequestService;
+
+  @PostMapping("/pull-request")
+  public ResponseEntity<ResponseDto> pullRequest(@RequestBody MemorialPullRequestDto dto) {
+    memorialPullRequestService.createMemorialPullRequest(dto);
+    return ResponseEntity.status(201).body(new ResponseDto("Memorial Commit is successfully requested to pull.", null));
+  }
+
+}
